feat (multer-config): modifies the multer-config file to handle sharp image processing and moved to the middlewares folder Sharp processing now only changes quality and format of the image
chore (auth.js): changes path to user model and
moved to middlewares folder

feat (app.js): - applies the changes made in
multer-config into the app.js file

- Sharp image processing is not handled anymore
in the routes but in the multer-config file

- Corrects a problem where single fields modifications
were not being saved in the database.
Now it works as expected.

chore(coomon.js): deletes a console.log that was
used for debugging and not needed anymore